% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{china_io_2020_153_df}
\alias{china_io_2020_153_df}
\title{Input-output Table for China, 2020 (153 Sectors)}
\format{
A data frame with 160 observations and 169 variables:
\describe{
  \item{Code}{Sector code (character)}
  \item{Description}{Sector description in English (character)}
  \item{DescriptionInChinese}{Sector description in Chinese (character)}
  \item{001}{Input from sector 001 (numeric)}
  \item{002}{Input from sector 002 (numeric)}
  \item{003}{Input from sector 003 (numeric)}
  \item{004}{Input from sector 004 (numeric)}
  \item{005}{Input from sector 005 (numeric)}
  \item{006}{Input from sector 006 (numeric)}
  \item{007}{Input from sector 007 (numeric)}
  \item{008}{Input from sector 008 (numeric)}
  \item{009}{Input from sector 009 (numeric)}
  \item{010}{Input from sector 010 (numeric)}
  \item{011}{Input from sector 011 (numeric)}
  \item{012}{Input from sector 012 (numeric)}
  \item{013}{Input from sector 013 (numeric)}
  \item{014}{Input from sector 014 (numeric)}
  \item{015}{Input from sector 015 (numeric)}
  \item{016}{Input from sector 016 (numeric)}
  \item{017}{Input from sector 017 (numeric)}
  \item{018}{Input from sector 018 (numeric)}
  \item{019}{Input from sector 019 (numeric)}
  \item{020}{Input from sector 020 (numeric)}
  \item{021}{Input from sector 021 (numeric)}
  \item{022}{Input from sector 022 (numeric)}
  \item{023}{Input from sector 023 (numeric)}
  \item{024}{Input from sector 024 (numeric)}
  \item{025}{Input from sector 025 (numeric)}
  \item{026}{Input from sector 026 (numeric)}
  \item{027}{Input from sector 027 (numeric)}
  \item{028}{Input from sector 028 (numeric)}
  \item{029}{Input from sector 029 (numeric)}
  \item{030}{Input from sector 030 (numeric)}
  \item{031}{Input from sector 031 (numeric)}
  \item{032}{Input from sector 032 (numeric)}
  \item{033}{Input from sector 033 (numeric)}
  \item{034}{Input from sector 034 (numeric)}
  \item{035}{Input from sector 035 (numeric)}
  \item{036}{Input from sector 036 (numeric)}
  \item{037}{Input from sector 037 (numeric)}
  \item{038}{Input from sector 038 (numeric)}
  \item{039}{Input from sector 039 (numeric)}
  \item{040}{Input from sector 040 (numeric)}
  \item{041}{Input from sector 041 (numeric)}
  \item{042}{Input from sector 042 (numeric)}
  \item{043}{Input from sector 043 (numeric)}
  \item{044}{Input from sector 044 (numeric)}
  \item{045}{Input from sector 045 (numeric)}
  \item{046}{Input from sector 046 (numeric)}
  \item{047}{Input from sector 047 (numeric)}
  \item{048}{Input from sector 048 (numeric)}
  \item{049}{Input from sector 049 (numeric)}
  \item{050}{Input from sector 050 (numeric)}
  \item{051}{Input from sector 051 (numeric)}
  \item{052}{Input from sector 052 (numeric)}
  \item{053}{Input from sector 053 (numeric)}
  \item{054}{Input from sector 054 (numeric)}
  \item{055}{Input from sector 055 (numeric)}
  \item{056}{Input from sector 056 (numeric)}
  \item{057}{Input from sector 057 (numeric)}
  \item{058}{Input from sector 058 (numeric)}
  \item{059}{Input from sector 059 (numeric)}
  \item{060}{Input from sector 060 (numeric)}
  \item{061}{Input from sector 061 (numeric)}
  \item{062}{Input from sector 062 (numeric)}
  \item{063}{Input from sector 063 (numeric)}
  \item{064}{Input from sector 064 (numeric)}
  \item{065}{Input from sector 065 (numeric)}
  \item{066}{Input from sector 066 (numeric)}
  \item{067}{Input from sector 067 (numeric)}
  \item{068}{Input from sector 068 (numeric)}
  \item{069}{Input from sector 069 (numeric)}
  \item{070}{Input from sector 070 (numeric)}
  \item{071}{Input from sector 071 (numeric)}
  \item{072}{Input from sector 072 (numeric)}
  \item{073}{Input from sector 073 (numeric)}
  \item{074}{Input from sector 074 (numeric)}
  \item{075}{Input from sector 075 (numeric)}
  \item{076}{Input from sector 076 (numeric)}
  \item{077}{Input from sector 077 (numeric)}
  \item{078}{Input from sector 078 (numeric)}
  \item{079}{Input from sector 079 (numeric)}
  \item{080}{Input from sector 080 (numeric)}
  \item{081}{Input from sector 081 (numeric)}
  \item{082}{Input from sector 082 (numeric)}
  \item{083}{Input from sector 083 (numeric)}
  \item{084}{Input from sector 084 (numeric)}
  \item{085}{Input from sector 085 (numeric)}
  \item{086}{Input from sector 086 (numeric)}
  \item{087}{Input from sector 087 (numeric)}
  \item{088}{Input from sector 088 (numeric)}
  \item{089}{Input from sector 089 (numeric)}
  \item{090}{Input from sector 090 (numeric)}
  \item{091}{Input from sector 091 (numeric)}
  \item{092}{Input from sector 092 (numeric)}
  \item{093}{Input from sector 093 (numeric)}
  \item{094}{Input from sector 094 (numeric)}
  \item{095}{Input from sector 095 (numeric)}
  \item{096}{Input from sector 096 (numeric)}
  \item{097}{Input from sector 097 (numeric)}
  \item{098}{Input from sector 098 (numeric)}
  \item{099}{Input from sector 099 (numeric)}
  \item{100}{Input from sector 100 (numeric)}
  \item{101}{Input from sector 101 (numeric)}
  \item{102}{Input from sector 102 (numeric)}
  \item{103}{Input from sector 103 (numeric)}
  \item{104}{Input from sector 104 (numeric)}
  \item{105}{Input from sector 105 (numeric)}
  \item{106}{Input from sector 106 (numeric)}
  \item{107}{Input from sector 107 (numeric)}
  \item{108}{Input from sector 108 (numeric)}
  \item{109}{Input from sector 109 (numeric)}
  \item{110}{Input from sector 110 (numeric)}
  \item{111}{Input from sector 111 (numeric)}
  \item{112}{Input from sector 112 (numeric)}
  \item{113}{Input from sector 113 (numeric)}
  \item{114}{Input from sector 114 (numeric)}
  \item{115}{Input from sector 115 (numeric)}
  \item{116}{Input from sector 116 (numeric)}
  \item{117}{Input from sector 117 (numeric)}
  \item{118}{Input from sector 118 (numeric)}
  \item{119}{Input from sector 119 (numeric)}
  \item{120}{Input from sector 120 (numeric)}
  \item{121}{Input from sector 121 (numeric)}
  \item{122}{Input from sector 122 (numeric)}
  \item{123}{Input from sector 123 (numeric)}
  \item{124}{Input from sector 124 (numeric)}
  \item{125}{Input from sector 125 (numeric)}
  \item{126}{Input from sector 126 (numeric)}
  \item{127}{Input from sector 127 (numeric)}
  \item{128}{Input from sector 128 (numeric)}
  \item{129}{Input from sector 129 (numeric)}
  \item{130}{Input from sector 130 (numeric)}
  \item{131}{Input from sector 131 (numeric)}
  \item{132}{Input from sector 132 (numeric)}
  \item{133}{Input from sector 133 (numeric)}
  \item{134}{Input from sector 134 (numeric)}
  \item{135}{Input from sector 135 (numeric)}
  \item{136}{Input from sector 136 (numeric)}
  \item{137}{Input from sector 137 (numeric)}
  \item{138}{Input from sector 138 (numeric)}
  \item{139}{Input from sector 139 (numeric)}
  \item{140}{Input from sector 140 (numeric)}
  \item{141}{Input from sector 141 (numeric)}
  \item{142}{Input from sector 142 (numeric)}
  \item{143}{Input from sector 143 (numeric)}
  \item{144}{Input from sector 144 (numeric)}
  \item{145}{Input from sector 145 (numeric)}
  \item{146}{Input from sector 146 (numeric)}
  \item{147}{Input from sector 147 (numeric)}
  \item{148}{Input from sector 148 (numeric)}
  \item{149}{Input from sector 149 (numeric)}
  \item{150}{Input from sector 150 (numeric)}
  \item{151}{Input from sector 151 (numeric)}
  \item{152}{Input from sector 152 (numeric)}
  \item{153}{Input from sector 153 (numeric)}
  \item{TIU}{Total intermediate use (numeric)}
  \item{FU101}{Final use category 101 (numeric)}
  \item{FU102}{Final use category 102 (numeric)}
  \item{THC}{Household consumption (numeric)}
  \item{FU103}{Final use category 103 (numeric)}
  \item{TC}{Total consumption (numeric)}
  \item{FU201}{Final use category 201 (numeric)}
  \item{FU202}{Final use category 202 (numeric)}
  \item{GCF}{Gross capital formation (numeric)}
  \item{EX}{Exports (numeric)}
  \item{TFU}{Total final use (numeric)}
  \item{IM}{Imports (numeric)}
  \item{GO}{Gross output (numeric)}
}
}
\source{
Data taken from the ionet package version 0.2.2
}
\usage{
data(china_io_2020_153_df)
}
\description{
This dataset, china_io_2020_153_df, is a data frame that represents the national input-output table of China for the year 2020.
It covers 153 sectors and provides inter-sectoral flows of goods and services. Data values are measured in 10,000 Chinese Yuan (CNY)
and calculated at producers' prices.
}
\details{
The dataset name has been kept as 'china_io_2020_153_df' to avoid confusion with other datasets in the R ecosystem.
This naming convention helps distinguish this dataset as part of the ChinAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a base R data frame. The original content has not been modified
in any way.
}
