% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinapis-package.R
\docType{package}
\name{ChinAPIs}
\alias{ChinAPIs}
\alias{ChinAPIs-package}
\title{ChinAPIs: Access Chinese Data via APIs and Curated Datasets}
\description{
{
This package provides functions to access data from public RESTful APIs including 'Nager.Date', 'World Bank API', and 'REST Countries API', retrieving real-time or historical data related to China, such as holidays, economic indicators, and international demographic and geopolitical indicators. Additionally, the package includes one of the largest curated collections of datasets focused on China and Hong Kong.
}
}
\details{
ChinAPIs: Access Chinese Data via APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access Chinese Data via APIs and Curated Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/chinapis}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
