% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{COVID19_HongKong_df}
\alias{COVID19_HongKong_df}
\title{COVID-19 Offspring Cases in Hong Kong (Jan–Apr 2020)}
\format{
A data frame with 290 observations and 2 variables:
\describe{
  \item{obs}{Number of offspring cases from a single seed case (numeric)}
  \item{type}{Type of transmission event (character)}
}
}
\source{
Data taken from the modelSSE package version 0.1-3
}
\usage{
data(COVID19_HongKong_df)
}
\description{
This dataset, COVID19_HongKong_df, is a data frame containing data on 290 observations of
offspring case numbers generated by individual seed cases during the COVID-19 outbreak
in Hong Kong, China, from January to April 2020. It includes the number of offspring
cases per seed and the type of transmission event.
}
\details{
The dataset name has been kept as 'COVID19_HongKong_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
ChinAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
