% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{hk_street_names_tbl_df}
\alias{hk_street_names_tbl_df}
\title{Hong Kong Street Names as of 2020}
\format{
A tibble with 4,603 observations and 21 variables:
\describe{
  \item{DC}{District code or abbreviation (character)}
  \item{StreetNames_EN}{Street name in English (character)}
  \item{StreetNames_ZH}{Street name in Chinese (character)}
  \item{TM}{Tuen Mun district indicator (logical)}
  \item{ST}{Sha Tin district indicator (logical)}
  \item{E}{Eastern district indicator (logical)}
  \item{S}{Southern district indicator (logical)}
  \item{WC}{Wan Chai district indicator (logical)}
  \item{C&W}{Central and Western district indicator (logical)}
  \item{Is}{Islands district indicator (logical)}
  \item{YL}{Yuen Long district indicator (logical)}
  \item{SK}{Sai Kung district indicator (logical)}
  \item{KC}{Kowloon City district indicator (logical)}
  \item{YTM}{Yau Tsim Mong district indicator (logical)}
  \item{KT}{Kwun Tong district indicator (logical)}
  \item{SSP}{Sham Shui Po district indicator (logical)}
  \item{N}{North district indicator (logical)}
  \item{TP}{Tai Po district indicator (logical)}
  \item{K&T}{Kwai Tsing district indicator (logical)}
  \item{TW}{Tsuen Wan district indicator (logical)}
  \item{WTS}{Wong Tai Sin district indicator (logical)}
}
}
\source{
Data taken from the hkdatasets package version 1.0.0
}
\usage{
data(hk_street_names_tbl_df)
}
\description{
This dataset, hk_street_names_tbl_df, is a tibble containing street names in Hong Kong as of the year 2020.
It includes English and Chinese names for each street and logical indicators of whether a street is located
within one of the 18 administrative districts of Hong Kong. This dataset is useful for geographic,
linguistic, and administrative analysis.
}
\details{
The dataset name has been kept as 'hk_street_names_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
ChinAPIs package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble (a modern form of data frame). The original content has not been modified
in any way.
}
