% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{top100name_year_df}
\alias{top100name_year_df}
\title{Top 100 Given Names in 6 Birth Cohorts}
\format{
A data frame with 100 observations and 37 variables:
\describe{
  \item{top100}{Ranking from 1 to 100 (numeric)}
  \item{name.all.1950}{Most common name (all genders) in 1950 (character)}
  \item{name.all.1960}{Most common name (all genders) in 1960 (character)}
  \item{name.all.1970}{Most common name (all genders) in 1970 (character)}
  \item{name.all.1980}{Most common name (all genders) in 1980 (character)}
  \item{name.all.1990}{Most common name (all genders) in 1990 (character)}
  \item{name.all.2000}{Most common name (all genders) in 2000 (character)}
  \item{n.all.1950}{Number of people with the name in 1950 (numeric)}
  \item{n.all.1960}{Number of people with the name in 1960 (numeric)}
  \item{n.all.1970}{Number of people with the name in 1970 (numeric)}
  \item{n.all.1980}{Number of people with the name in 1980 (numeric)}
  \item{n.all.1990}{Number of people with the name in 1990 (numeric)}
  \item{n.all.2000}{Number of people with the name in 2000 (numeric)}
  \item{name.m.1950}{Most common male name in 1950 (character)}
  \item{name.m.1960}{Most common male name in 1960 (character)}
  \item{name.m.1970}{Most common male name in 1970 (character)}
  \item{name.m.1980}{Most common male name in 1980 (character)}
  \item{name.m.1990}{Most common male name in 1990 (character)}
  \item{name.m.2000}{Most common male name in 2000 (character)}
  \item{n.m.1950}{Number of males with the name in 1950 (numeric)}
  \item{n.m.1960}{Number of males with the name in 1960 (numeric)}
  \item{n.m.1970}{Number of males with the name in 1970 (numeric)}
  \item{n.m.1980}{Number of males with the name in 1980 (numeric)}
  \item{n.m.1990}{Number of males with the name in 1990 (numeric)}
  \item{n.m.2000}{Number of males with the name in 2000 (numeric)}
  \item{name.f.1950}{Most common female name in 1950 (character)}
  \item{name.f.1960}{Most common female name in 1960 (character)}
  \item{name.f.1970}{Most common female name in 1970 (character)}
  \item{name.f.1980}{Most common female name in 1980 (character)}
  \item{name.f.1990}{Most common female name in 1990 (character)}
  \item{name.f.2000}{Most common female name in 2000 (character)}
  \item{n.f.1950}{Number of females with the name in 1950 (numeric)}
  \item{n.f.1960}{Number of females with the name in 1960 (numeric)}
  \item{n.f.1970}{Number of females with the name in 1970 (numeric)}
  \item{n.f.1980}{Number of females with the name in 1980 (numeric)}
  \item{n.f.1990}{Number of females with the name in 1990 (numeric)}
  \item{n.f.2000}{Number of females with the name in 2000 (numeric)}
}
}
\source{
Data taken from the ChineseNames package version 2023.8
}
\usage{
data(top100name_year_df)
}
\description{
This dataset, top100name_year_df, is a data frame containing the top 100 given names in China across
six birth cohorts: 1950, 1960, 1970, 1980, 1990, and 2000. It includes rankings and frequencies
for all individuals, as well as separately for males and females. The dataset provides insights into
naming trends and gender differences over time.
}
\details{
The dataset name has been kept as 'top100name_year_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
ChinAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
